RIL: Squash patches for bringing back registerForOn
Adjust for XOS 7.1

Revert "Use radio availability to sync card status instead of radio on"

 * Revert to AOSP behaviour for this. Introduce a few fixes later on
   which fix radio availability issues on both QC and legacy devices.

This reverts commit af1cd8cbac506462e91d024f9c7c864d6002495b.

Change-Id: Iae4063337d13eabc979acd172133d7744a270647

UiccController: use registerForAvailable only for persist.radio.apm_sim_not_pwdn

Some QCOM RILs allow the SIM to be loaded in airplane mode, thus
CAF changed the callback to registerForAvailable, but that does
not work on other RILs where the SIM information will not be
loaded successfully.

Change-Id: I9f9124417db2fe983e716a2825fc08d87511f8aa

Use registerForAvailable when device is encrypted only for RIL version >= 9

On older RIL versions, registerForAvailable causes the sim to not
being detected. We fixed this for non-encrypted state, but missed it
for encrypted state.

Change-Id: I26cca33c10edcbcb8ea10daaa6104eed910354bd
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>