PhoneFactory: Don't force a preferred mode if the default includes TD-SCDMA
After getting the default network mode from ro.telephony.default_network
(via RILConstants), we were ignoring it if the phone claimed LTE support
and overriding with either GLOBAL (for CDMA+LTE) or LTE_GSM_WCDMA (for
GSM). This is of arguable benefit to begin with, and it totally messes
things if the original mode has TD-SCDMA, which ends up entirely disabled
by default. So for now, skip the overrides if the default includes
TD-SCDMA.
And in the future, we should get rid of this entirely. If a phone
ships mis-configured and doesn't set the correct ro.telephony.default_network,
tough luck. It's not the radio stack's job to compensate for poor
default configurations.

Change-Id: Ie00fcf03977f6048c96bad2579f505804ed6628d
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>